requirements.txt
Django==3.0.5
djangorestframework==3.11.1
django-rest-swagger==2.2.0
PyYAML==5.3.1
python-decouple==3.3
djongo==1.3.3
django-filter==2.3.0